Day 9/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
beautifulsoup4>=4.12.0
gradio==4.19.2
python-dotenv>=1.0.0
chromadb==0.4.24
html5lib>=1.1
google-generativeai>=0.3.0
pydantic>=2.0.0
typing-extensions>=4.5.0
faiss-cpu>=1.7.4  # For vector storage
tiktoken>=0.5.0  # For token counting 